parser: Parse "Depends-on" tags in emails
Add a new function to parse "Depends-on" tags to the parser. The value
may either be the message ID of a patch or cover letter email already
received by Patchwork, or the web URL of a patch or series. When this
tag is found, the parser will add the series (or the series the patch
belongs to) as a dependency to the series it is creating.

Signed-off-by: Adam Hassick <ahassick@iol.unh.edu>